Refactor config setup
