+++ database/connect.ts
@@ -1,6 +1,6 @@
 import 'server-only';
 import postgres from 'postgres';
-import { setEnvironmentVariables } from '../ley.config.mjs';
+import { setEnvironmentVariables } from '../util/config.mjs';
 
 // This loads all environment variables from a .env file
 // for all code after this line
